requirements.txt
# Core RAG Framework
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38

# Vector Database
pinecone-client==3.2.2
pinecone-text==0.7.1

# Embeddings and LLMs
openai==1.25.0
tiktoken==0.6.0

# Evaluation
ragas==0.1.7
datasets==2.18.0

# Data Processing
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1

# Chunking and Text Processing
beautifulsoup4==4.12.3
markdown==3.5.2
pypdf==4.1.0
python-magic==0.4.27

# Observability
langsmith==0.1.52
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
prometheus-client==0.20.0

# API and Web
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1

# Utilities
tqdm==4.66.2
jsonlines==4.0.0
pyyaml==6.0.1

# Development
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
black==24.3.0
ruff==0.3.5
